LayoutTests/platform/qt/touchadjustment/disabled-formelements-expected.txt

Touch Adjustment : Disabled form elements - bug 92093

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test we can hit the form elements directly.
PASS adjusted node was null.
PASS adjusted node was LABEL#label2.
FAIL adjusted node should be null. Was INPUT#input1.
PASS adjusted node was INPUT#input2.
Test indirect touches are adjusted to the enabled form elements.
PASS adjusted node was LABEL#label2.
PASS adjusted node was LABEL#label2.
PASS adjusted node was LABEL#label2.
PASS adjusted node was INPUT#input2.
FAIL adjusted node should be INPUT#input2. Was INPUT#input1.
FAIL adjusted node should be INPUT#input2. Was INPUT#input1.
PASS successfullyParsed is true

TEST COMPLETE
